# fix(engine): Improved query-solver performance