--- a/crates/engine/query-solver/src/solve/steiner_tree/mod.rs
+++ b/crates/engine/query-solver/src/solve/steiner_tree/mod.rs
@@ -1,7 +1,7 @@
-mod graph;
+mod context;
 mod shortest_path;
 #[cfg(test)]
 mod tests;
 
-use graph::*;
+pub(crate) use context::*;
 pub(crate) use shortest_path::*;