… Cargo.lock
… crates/engine/query-solver/src/solve/steiner_tree/context.rs
… crates/engine/query-solver/src/solve/steiner_tree/shortest_path.rs
… crates/engine/query-solver/src/solve/steiner_tree/shortest_path/tree.rs
… crates/engine/query-solver/src/tests/snapshots/engine_query_solver__tests__typename__only_typename-solved.snap
… crates/engine/query-solver/src/tests/snapshots/engine_query_solver__tests__typename__only_typename-solver.snap
… crates/engine/query-solver/src/tests/snapshots/engine_query_solver__tests__typename__top_level_typename-solved.snap
… crates/engine/query-solver/src/tests/snapshots/engine_query_solver__tests__typename__top_level_typename-solver.snap
… crates/engine/src/prepare/operation/after_variables.rs
… crates/engine/src/prepare/operation/before_variables.rs
… crates/engine/src/prepare/operation/mod.rs
… crates/grafbase-workspace-hack/Cargo.toml